+++ mods/BrewLAN_Units/BrewTeaParty/hook/lua/ui/help/unitdescription.lua
@@ -1,0 +1,2 @@
+--Experimental Dread
+Description['sas0401'] = "<LOC Unit_Description_BrewTeaParty_001>Experimental dreadnought-type battlecruiser. Features light armor for increased movement speed, with personal shields and anti-torpedo weapons for defences. Offensively armed with Oblivion cannons, shield disruption cannons, anti-air missiles, and a secondary fragmentation artillery."