+++ mods/BrewLAN_Units/BrewTeaParty/mod_info.lua
@@ -1,0 +1,16 @@
+name = "BrewLAN: Tea Party"
+uid = "25D57D85-9JA7-D842-BREW-BOATS00000000"
+version = 0
+copyright = "© 2020–2020 Sean Wheeldon"
+description = "Adds an Aeon Experimental Battlecruiser."
+author = "Balthazar"
+--icon = "/mods/BrewLAN_Units/BrewShields/mod_icon.png"
+selectable = true
+enabled = true
+exclusive = false
+ui_only = false
+requires = { }
+requiresNames = { }
+conflicts = { }
+before = { }
+after = { }